Create new events
* DatabaseTable allows for fluent SQL statement generation
* Including font-awesome to avoid stalled loading
* loading-overlay to block input during loading
* JDBC4 primary key generation
* Added lombok

+++ src/main/java/net/openright/conferencer/domain/orders/Order.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 class Order {
 
     private String title;
-    private Integer id;
+    private Long id;
     private List<OrderLine> orderLines = new ArrayList<>();
 
     Order(String title) {
@@ -27,11 +27,11 @@ class Order {
         return this;
     }
 
-    void setId(int id) {
+    void setId(long id) {
         this.id = id;
     }
 
-    int getId() {
+    long getId() {
         return id;
     }
